(GH-3) Determine type automatically if not specified
With 4aa930627, checksum always determined the type of checksum type to
use based on the length of the checksum. However that should only
happen if checksum type is not passed. If checksum type is given, it
should always use that type, even if it is incorrect and the hash would
have been correct using another type.