(#11) Add Cake scripts and configuration
This commit adds the recipe.cake entry point for the
Chocolatey.Cake.Recipe package, and configures it to be specific to the
checksum project. This includes selecting which files need to be signed,
and obfuscated, etc.

It also adds the entry point files, i.e. build.bat, build.official.bat,
etc.