(#16) Add VERIFICATION.txt to package
The package-valdiator on Chocolatey Commnity Repository now requires
that any embedded package contain a VERIFICATION.txt file, explaining
where a person can find information about the contents of the package,
namely a SHA of the binaries contained within it.

This wasn't required the last time a version of this package was pushed
to CCR, so the push of .0.3.0 failed, since it didn't include it.